Early Power Flow
Battery seems to charge and discharge, needs improved logic for battery handling and the speed at which turbines slow down to if there's too much power.  Maybe add logic to turn one/more off, or calculate the speed required to be spun at in order to generate the necessary power.